+++ CHANGELOG.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Updated parameter help to correctly suggest `local` to be used
 * Added additional examples that describe usage of the -PrimaryClusterID parameter
 
+### Changed [Get-RubrikReportData]
+
+* Default behavior of Get-RubrikReportData updated to reflect default behavior of other parameters, setting limit to maximum amount unless specified. [Issue 408](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/408)
+
 ## 2019-07-30
 
 ### Changed [readme.md]